requirements.txt
antlr4-python3-runtime==4.9.1

graphviz~=0.16